requirements.txt
openai>=1.35
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
plotly>=5.20.0
streamlit>=1.34.0
python-dotenv>=1.0.1
yandex-cloud-ml-sdk>=0.1.6
pyjwt[crypto]>=2.9.0
requests>=2.31.0
tqdm>=4.66.0
jinja2>=3.1.3
tenacity>=8.2.3
altair>=5.3.0
pyarrow>=15.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
telethon>=1.34.0

